Fixed getting success msg on failing swift operation
Show error message instead of success message when attempting to
create a duplicate object or pseudo-folder. Added tests for
creating and creating duplicate pseudo-folders.

Closes-Bug: #1478126

+++ openstack_dashboard/api/swift.py
@@ -277,6 +277,8 @@ def swift_copy_object(request, orig_container_name, orig_object_name,
 
 def swift_upload_object(request, container_name, object_name,
                         object_file=None):
+    if swift_object_exists(request, container_name, object_name):
+        raise exceptions.AlreadyExists(object_name, 'object')
     headers = {}
     size = 0
     if object_file:
@@ -294,6 +296,10 @@ def swift_upload_object(request, container_name, object_name,
 
 
 def swift_create_pseudo_folder(request, container_name, pseudo_folder_name):
+    # Make sure the folder name doesn't already exist.
+    if swift_object_exists(request, container_name, pseudo_folder_name):
+        name = pseudo_folder_name.strip('/')
+        raise exceptions.AlreadyExists(name, 'pseudo-folder')
     headers = {}
     etag = swift_api(request).put_object(container_name,
                                          pseudo_folder_name,
